Model loader method used by CoC-StellarEngine
Reupload but in a folder this time. Mb

Uploading the ModelManager and Model headers&source files that I used for 'Cabinet of Curiosities' in the 'Stellar Engine'.

The method entails browsing through all the files in the model folder and creating a json file that contains an array of 2 values. This way the developers don't need to pass on the full file path + name and create extra data all over the code. Now instead there is one modelmanager that loads all the models once which then can be used to retrieve the model data just by using the model's filename.